requirements.txt
bjoern
falcon
deepspeech-gpu